Create youtube_tools.py
